+++ CHANGELOG.md
@@ -1,6 +1,13 @@
 warc2zim
 ===
 
+# 1.4.0
+
+* Additional fuzzy matching rules for youtube and vimeo, and additional test cases
+* Support for youtube videos, which require POST request handling to work.
+* Support for canonicalizing POST request data into URL for fuzzy matching (using cdxj-indexer)
+* Support loading custom sw.js from a local file path
+
 # 1.3.6
 
 * updated zimscraperlib (for libzim fix)